ProyectoWeb/requirements.txt
django == 3.1.2
pillow
django-crispy-forms
django-cleanup
django-rest-framework
django-admin-interface

#pip install -r requirements.txt